my second commit

+++ test_factorial.java
@@ -1,7 +1,7 @@
 public class test_factorial{
     public static void main(String[] args){
         int n=5;
-        int result=Factorial.calculate(n);
+        long result=Factorial.calculate(n);
         System.out.println("result is"+result);
     }
 }
